fix: update integration tests

+++ packages/gql/src/resolver/collection.resolver.ts
@@ -89,10 +89,13 @@ const removeCollectionDuplicates = async (
               }
             }),
           )
-          const removeIds = toRemove.map((collection) => collection.id)
-          await repositories.collection.hardDeleteByIds(removeIds)
-          await seService.deleteCollections(toRemove)
-          removedDuplicates = true
+
+          if (toRemove.length) {
+            const removeIds = toRemove.map((collection) => collection.id)
+            await repositories.collection.hardDeleteByIds(removeIds)
+            await seService.deleteCollections(toRemove)
+            removedDuplicates = true
+          }
         }
       }),
     )
